Support OpenCL implementations that don't support structs.
Defining PYOPENCL_COMPLEX_USE_VECTOR will fall back to vector-based implementation for problematic platforms.  To support this, all references to a.real and a.imag have been replaced with TPROOT##_real(a) and TPROOT##_imag(a) for rvalues, and a.PYOPENCL_COMPLEX_REAL and a.PYOPENCL_COMPLEX_IMAG for lvalues.